requirements.txt
absl-py==0.9.0
aiohttp==3.7.4.post0
astor==0.8.1
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
Backtesting==0.3.1
backtrader==1.9.76.123
beautifulsoup4==4.9.3
bokeh==2.3.1
boto3==1.17.8
botocore==1.20.8
bs4==0.0.1
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
dependency-injector==4.20.2
discord.py==1.7.2
dnspython==1.16.0
docker==4.4.4
Flask==1.1.2
Flask-PyMongo==2.3.0
gast==0.2.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.31.0
h5py==2.10.0
httplib2==0.19.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.4.0
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
json5==0.9.5
kaggle==1.5.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lean==0.1.52
lxml==4.6.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
memory-profiler==0.57.0
morss==0.0.0
multidict==5.1.0
networkx==2.4
numpy==1.19.1
opencv-python==4.4.0.40
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.1
Pillow==7.2.0
protobuf==3.12.4
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pydantic==1.7.3
Pygments==2.9.0
pylint==2.6.0
pymongo==3.11.0
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-slugify==4.0.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
quantconnect-stubs==11594
requests==2.25.1
rich==9.10.0
rsa==4.7.1
s3transfer==0.3.4
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
sgfmill==1.1.1
Shapely==1.7.0
six==1.15.0
soupsieve==2.2
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.8.13
toml==0.10.1
tornado==6.1
tqdm==4.48.2
typed-ast==1.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.25.11
webcolors==1.11.1
websocket-client==0.59.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
